requirements.txt
kafka-python==2.0.2
protobuf==3.14.0
six==1.15.0
